requirements.txt
telebot==0.0.5
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-dateutil==2.9.0  # Oy hisoblash uchun
matplotlib==3.9.2  # Grafiklar
pandas==2.2.3  # Statistika uchun (optional, lekin yaxshi)
logging  # Built-in
psycopg2-binary==2.9.9  # PostgreSQL uchun (agar o'tkazsangiz)
psycopg2-binary